バナーをチームバナーに拡張
